Complete code overhaul (lot of work still to do)
