+++ src/app/app-routing.module.ts
@@ -1,19 +1,15 @@
 import { NgModule } from "@angular/core";
 import { RouterModule, Routes } from "@angular/router";
 
-import { EHREntryComponent } from "./ehr-entry/ehr-entry.component";
-import { DataEntryFormComponent } from "./data-entry/data-entry.component";
 import { SMARTLaunchComponent } from "./smart-initialization/smart-launch.component";
 import { SMARTTokenReceptionComponent } from "./smart-initialization/smart-token-reception.component";
-import { VisualizeResultsComponent } from "./visualize-results/visualize-results.component";
+import {PrimaryFunctionalityComponent} from "./primary-functionality/primary-functionality.component";
 
 const routes: Routes = [
-  { path: "", redirectTo: "/data-entry", pathMatch: "full" },
-  { path: "ehr-entry", component: EHREntryComponent },
-  { path: "data-entry", component: DataEntryFormComponent },
+  { path: "", redirectTo: "/primary", pathMatch: "full" },
   { path: "smart-launch", component: SMARTLaunchComponent },
   { path: "token-reception", component: SMARTTokenReceptionComponent },
-  { path: "visualize-results", component: VisualizeResultsComponent },
+  { path: "primary", component: PrimaryFunctionalityComponent }
 ];
 
 @NgModule({
