+++ src/app/app.component.ts
@@ -3,20 +3,10 @@ import { Component } from "@angular/core";
 @Component({
   selector: "app-root",
   template: `
-    <info-header></info-header>
-    <div id="content">
-      <router-outlet></router-outlet>
-    </div>
+    <header></header>
+    <router-outlet></router-outlet>
   `,
-  styles: [`    
-    #content {
-      display: block;
-      margin: 0 20px;
-      width: calc(100% - 40px);
-    }
-  `]
 })
-
 export class AppComponent {
-  title = "SMART EHR Module (Angular)";
+  title = "SMART-CO Application";
 }
