+++ src/app/smart-initialization/smart-token-reception.component.ts
@@ -25,6 +25,6 @@ export class SMARTTokenReceptionComponent implements OnInit {
     this.smartReferenceService.ready();
 
     // Redirect to the disease selection.
-    setTimeout(() => this.router.navigate(["/data-entry"]), 100);
+    setTimeout(() => this.router.navigate(["/primary"]), 100);
   }
 }
